feat(actioner): move actioner to a seperate class for easier structuring
